### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-21024

### Test account
- environment:UAT
- username: AD-Account
- salesOrg:2001
- customer:30082707
- shipto:70149863
- material/setup:

### Output screenshot
-[Screen Recording 2024-04-17 at 12.55.20 PM.mov](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/102604/attachments/Screen%20Recording%202024-04-17%20at%2012.55.20%20PM.mov)

EZRX-21024 Mobile v3: Add validation to disable download button once download begins and change the text to Open Payment Advice once downloaded

+++ lib/presentation/payments/payment_summary_details/widgets/payment_advice_button.dart
@@ -17,7 +17,7 @@ class _PaymentAdviceButton extends StatelessWidget {
             final eligibilityState = context.read<EligibilityBloc>().state;
             if (!state.isSavingAdvice && !eligibilityState.salesOrg.isID) {
               CustomSnackBar(
-                messageText: context.tr('Download Successful'),
+                messageText: context.tr('Payment advice has been downloaded.'),
               ).show(context);
             }
             if (!state.isCancelingAdvice && eligibilityState.salesOrg.isID) {
@@ -50,7 +50,8 @@ class _PaymentAdviceButton extends StatelessWidget {
           previous.isSavingAdvice != current.isSavingAdvice ||
           previous.isFetchingAdvice != current.isFetchingAdvice ||
           previous.isDeletingPayment != current.isDeletingPayment ||
-          previous.isCancelingAdvice != current.isCancelingAdvice,
+          previous.isCancelingAdvice != current.isCancelingAdvice ||
+          previous.savedAdvice != current.savedAdvice,
       builder: (context, state) {
         final eligibilityState = context.read<EligibilityBloc>().state;
         var buttons = _getDefaultButtons(state);
@@ -101,9 +102,8 @@ class _PaymentAdviceButton extends StatelessWidget {
     PaymentSummaryDetailsState state,
   ) =>
       [
-        if (!state.details.status.getIsSuccessfulOrProcessed)
-          _DeleteAdviceButton(state: state),
-        if (!state.details.status.getIsSuccessfulOrProcessed)
+        if (state.showDeleteAdviceButton) _DeleteAdviceButton(state: state),
+        if (state.showDeleteAdviceButton)
           const SizedBox(
             width: 16,
           ),
@@ -185,18 +185,22 @@ class _DownloadAdviceButton extends StatelessWidget {
           child: LoadingShimmer.withChild(
             enabled: state.isSavingAdvice,
             child: Text(
-              context.tr('Download advice'),
+              state.isSavedAdviceEmpty
+                  ? context.tr('Download advice')
+                  : context.tr('Open payment advice'),
               style: Theme.of(context).textTheme.labelMedium?.copyWith(
                     color: ZPColors.primary,
                   ),
             ),
           ),
         ),
-        onPressed: () {
-          context.read<PaymentSummaryDetailsBloc>().add(
-                const PaymentSummaryDetailsEvent.saveAdvice(),
-              );
-        },
+        onPressed: () => state.isSavedAdviceEmpty
+            ? context.read<PaymentSummaryDetailsBloc>().add(
+                  const PaymentSummaryDetailsEvent.saveAdvice(),
+                )
+            : context.read<PaymentSummaryDetailsBloc>().add(
+                  const PaymentSummaryDetailsEvent.viewSavedAdvice(),
+                ),
       ),
     );
   }
